Python 3 fixes: convert source to text
In Python 3 the SCSS source code must be decoded from binary to text
before processing. The latest compressor passes a charset argument to
our filter, which will be used. If not passed, it will default to utf-8.

Note that Python 3 support currently depends on the latest (master)
compressor
and a pyScss with this fix incorporated:
https://github.com/wojas/pyScss/commit/e4cf419509f3c568c108c3196e2fda50bcc37536

NOTE: Tests related to the sprites fail for me, both on python 2.7 and
3.4. The hash is different from what's expected.